# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DamascusTransit API",
    "description": "Real-time transit tracking and fleet management \u2014 multi-tenant SaaS edition.\n\n## Authentication\n\nMost read endpoints are public but require an `?operator=<slug>` query parameter.\n1. **POST /api/v1/auth/login** \u2014 exchange email/password for a JWT token\n2. Include the token as `Authorization: Bearer <token>` on protected endpoints\n\nRoles: `super_admin` \u00b7 `admin` \u00b7 `dispatcher` \u00b7 `driver` \u00b7 `viewer`\n\nTokens expire after 24 hours.",
    "contact": {
      "name": "Damascus Transit System",
      "email": "support@damascustransit.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/driver/position": {
      "post": {
        "tags": [
          "driver"
        ],
        "summary": "Report Driver Position",
        "description": "Report driver's current position.",
        "operationId": "report_driver_position_api_driver_position_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PositionUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusTimestampResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/driver/trip/start": {
      "post": {
        "tags": [
          "driver"
        ],
        "summary": "Start Trip",
        "description": "Start a new trip for the driver.",
        "operationId": "start_trip_api_driver_trip_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripStart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/driver/trip/end": {
      "post": {
        "tags": [
          "driver"
        ],
        "summary": "End Trip",
        "description": "End the driver's current trip.",
        "operationId": "end_trip_api_driver_trip_end_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripEnd"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TripActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/driver/trip/passenger-count": {
      "post": {
        "tags": [
          "driver"
        ],
        "summary": "Update Passenger Count",
        "description": "Update passenger count for current trip.",
        "operationId": "update_passenger_count_api_driver_trip_passenger_count_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PassengerCountUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusTimestampResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PassengerCountUpdate": {
        "properties": {
          "passenger_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Passenger Count"
          }
        },
        "type": "object",
        "required": [
          "passenger_count"
        ],
        "title": "PassengerCountUpdate"
      },
      "PositionUpdate": {
        "properties": {
          "latitude": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Longitude"
          },
          "speed_kmh": {
            "anyOf": [
              {
                "type": "number",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed Kmh"
          },
          "heading": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 360.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Heading"
          }
        },
        "type": "object",
        "required": [
          "latitude",
          "longitude"
        ],
        "title": "PositionUpdate"
      },
...
      "StatusTimestampResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "status",
          "timestamp"
        ],
        "title": "StatusTimestampResponse"
      },
...
      "TripActionResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "trip_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trip Id"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "status",
          "timestamp"
        ],
        "title": "TripActionResponse"
      },
      "TripEnd": {
        "properties": {
          "passenger_count": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Passenger Count"
          }
        },
        "type": "object",
        "title": "TripEnd"
      },
      "TripStart": {
        "properties": {
          "route_id": {
            "type": "string",
            "title": "Route Id"
          },
          "scheduled_departure": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled Departure"
          }
        },
        "type": "object",
        "required": [
          "route_id"
        ],
        "title": "TripStart"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}